+++ test/Test/Runtime/EndToEndDataMovement.hs
@@ -28,7 +28,7 @@ tests = testGroup "EndToEnd.DataMovement"
                 [ FuncArg "arg0" (TensorType [5] F32) ]
                 $ do
                     x <- arg @'[5] @'F32
-                    y <- slice x [1] [4] [1]
+                    y <- slice x (v1 1) (v1 4) (v1 1)
                     return y
         exec <- compile api client (render modu)
         let inp = V.fromList [0.0, 1.0, 2.0, 3.0, 4.0]
@@ -41,7 +41,7 @@ tests = testGroup "EndToEnd.DataMovement"
                 [ FuncArg "arg0" (TensorType [5] F32) ]
                 $ do
                     x <- arg @'[5] @'F32
-                    y <- slice x [0] [4] [2]
+                    y <- slice x (v1 0) (v1 4) (v1 2)
                     return y
         exec <- compile api client (render modu)
         let inp = V.fromList [0.0, 1.0, 2.0, 3.0, 4.0]
@@ -55,7 +55,7 @@ tests = testGroup "EndToEnd.DataMovement"
                 $ do
                     x <- arg @'[2] @'F32
                     padVal <- constant @'[] @'F32 0.0
-                    y <- pad x padVal [1] [1] [0]
+                    y <- pad x padVal (v1 1) (v1 1) (v1 0)
                     return y
         exec <- compile api client (render modu)
         let inp = V.fromList [1.0, 2.0]
@@ -200,7 +200,7 @@ tests = testGroup "EndToEnd.DataMovement"
                 $ do
                     x <- arg @'[4] @'F32
                     idx <- constant @'[] @'I64 1
-                    y <- dynamicSlice x [idx] [2]
+                    y <- dynamicSlice x [idx] (v1 2)
                     return y
         exec <- compile api client (render modu)
         let inp = V.fromList [0.0, 1.0, 2.0, 3.0]