+++ test/Test/Runtime/EndToEndShapeGPU.hs
@@ -44,7 +44,7 @@ tests getGPU = testGroup "EndToEnd.ShapeGPU"
                 [ FuncArg "arg0" (TensorType [2, 2] F32) ]
                 $ do
                     x <- arg
-                    y <- transpose @'[2, 2] @'[2, 2] [1, 0] x
+                    y <- transpose @'[2, 2] @'[2, 2] (v2 1 0) x
                     return y
         exec <- compile api client (render modu)
         bufIn <- toDeviceF32On api client dev input2x2 [2, 2]
@@ -57,7 +57,7 @@ tests getGPU = testGroup "EndToEnd.ShapeGPU"
                 [ FuncArg "arg0" (TensorType [2, 2] F32) ]
                 $ do
                     x <- arg
-                    y <- transpose @'[2, 2] @'[2, 2] [0, 1] x
+                    y <- transpose @'[2, 2] @'[2, 2] (v2 0 1) x
                     return y
         exec <- compile api client (render modu)
         bufIn <- toDeviceF32On api client dev input2x2 [2, 2]